% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{blogs}
\alias{blogs}
\title{Text Corpus Culled from Blogs}
\format{\preformatted{ chr [1:18125] "could be added to the gallery wall." ...
}}
\usage{
blogs
}
\description{
A text corpus containing 18,125 English language blog entries.  These entries
contain 52,172 sentences and 766,222 words.
}
\keyword{datasets}

