% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clean_sentences}
\alias{clean_sentences}
\title{Cleans sentences used as input to the model.}
\usage{
clean_sentences(sentences, start_tag = "^", end_tag = "$")
}
\arguments{
\item{sentences}{The sentences to pre-process.}

\item{start_tag}{The tag used to mark the beginning of a sentence.}

\item{end_tag}{The tag used to mark the end of a sentence.}
}
\value{
A vector of cleaned, pre-processed sentences.
}
\description{
Performs pre-processing on text input that has been already been split
into sentences.  Each element of the character vector input should
denote a single sentence.
}
\details{

}

