% Generated by roxygen2 (4.0.2): do not edit by hand
\name{except_last_word}
\alias{except_last_word}
\title{Excludes the last word from a phrase of text.}
\usage{
except_last_word(text)
}
\arguments{
\item{text}{The phrase containing words.}
}
\value{
A character vector containing 1 or more words.
}
\description{
Given a phrase of text the function will return all words except the
last word within the phrase.
}
\details{

}

