% Generated by roxygen2 (4.0.2): do not edit by hand
\name{last_word}
\alias{last_word}
\title{Finds the last word from a phrase of text.}
\usage{
last_word(text)
}
\arguments{
\item{text}{The phrase containing words.}
}
\value{
A character vector containing only the last word in the phrase.
}
\description{
Given a phrase of text the function will return the last word.
}
\details{

}

