% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{news}
\alias{news}
\title{Text Corpus Culled from News Sources}
\format{\preformatted{ chr [1:20529] "\\"The Voice,\\" NBC's upstart singing competition, is back for its second season Sunday, and the network is kicking it off in pr"| __truncated__ ...
}}
\usage{
news
}
\description{
A text corpus containing 20,529 English language news stories.  These contain
49,501 sentences and 704,148 words.
}
\keyword{datasets}

