% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ngram_mle}
\alias{ngram_mle}
\title{Calculate MLE of N-gram Probability}
\usage{
ngram_mle(ngrams)
}
\arguments{
\item{ngrams}{A data.table containing all contexts in the corpus.}
}
\value{
The maximum liklihood estimate of the probability of each n-gram.
}
\description{
Calculates the maximum liklihood estimate of the probability of any
N-gram based on the training corpus.
}
\details{
p (we love you) = count (we love you) / count (we love)
     p (we love)     = count (we love)     / count (we)
     p (we)          = count (we)          / count (<start-of-sentence>)
}

