% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.ngram}
\alias{predict.ngram}
\title{Prediction with N-Gram Language Models}
\usage{
\method{predict}{ngram}(object, newdata, rank = 5)
}
\arguments{
\item{object}{The N-gram model to use for prediction.}

\item{newdata}{A phrase to predict the next word for.}

\item{rank}{The top 'rank' most likely words are returned.}
}
\value{
The next mostly likely word(s).
}
\description{
Predicts the next most likely word using an N-Gram language model.
}
\details{

}
\examples{
\dontrun{
data (blogs)
fit <- ngram (blogs)
predict (fit, "What")
}
}

