% Generated by roxygen2 (4.0.2): do not edit by hand
\name{split_by_ngram}
\alias{split_by_ngram}
\title{Splits sentences into N-grams.}
\usage{
split_by_ngram(sentences, minN, maxN, delimiters)
}
\arguments{
\item{sentences}{The sentences to split into n-grams.}

\item{minN}{The minimum value of N defining the order of N-grams to create.}

\item{maxN}{The maximum value of N defining the order of N-grams to create.}

\item{delimiters}{The delimiters used to split the sentences into words.}
}
\value{
A data.table containing n-grams.
}
\description{
Splits a vector of sentences into N-grams.  The function accepts a range
of N values over which to create the N-grams.
}

