% Generated by roxygen2 (4.0.2): do not edit by hand
\name{split_by_sentence}
\alias{split_by_sentence}
\title{Splits text into sentences.}
\usage{
split_by_sentence(text)
}
\arguments{
\item{The}{raw text to split on sentence boundaries.}
}
\value{
A character vector containing 1 or more sentences.
}
\description{
Splits text into 1 or more sentences based on sentence boundaries. Sentence
boundaries are demarcated by 1 or more period (.), exclam (!), or question
mark (?).
}
\details{

}

