% Generated by roxygen2 (4.0.2): do not edit by hand
\name{split_by_word}
\alias{split_by_word}
\title{Splits text into words.}
\usage{
split_by_word(text)
}
\arguments{
\item{The}{raw text to split into words.}
}
\value{
A character vector containing 1 or more words.
}
\description{
Splits text into 1 or more words.
}
\details{

}

