% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{twitter}
\alias{twitter}
\title{Text Corpus Culled from Twitter}
\format{\preformatted{ chr [1:47390] "Good questions. RT : Your #brand will be judged based on its #website. Is your website a good brand ambassador?..." ...
}}
\usage{
twitter
}
\description{
A text corpus containing 47,390 tweets created on Twitter.  These contain 84,256
sentences and 613,412 words.
}
\keyword{datasets}

